Fixed improper exception handling of lsass dump parsing

+++ cme/modules/lsassy_dump.py
@@ -56,12 +56,18 @@ class CMEModule:
         if dumper is None:
             context.log.error("Unable to load dump method '{}'".format(self.method))
             return False
+
         file = dumper.dump()
         if file is None:
             context.log.error("Unable to dump lsass")
             return False
 
-        credentials, tickets = Parser(file).parse()
+        parsed = Parser(file).parse()
+        if parsed is None:
+            context.log.error("Unable to parse lsass dump")
+            return False
+        credentials, tickets = parsed
+
         file.close()
         ImpacketFile.delete(session, file.get_file_path())
         if credentials is None:
@@ -79,6 +85,7 @@ class CMEModule:
         if len(credentials) == 0:
             context.log.info("No credentials found")
         credz_bh = []
+        domain = None
         for cred in credentials:
             domain = cred["domain"]
             if "." not in cred["domain"] and cred["domain"].upper() in connection.domain.upper():
